requirements.txt
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
numpy==1.25.2
scipy==1.11.1
matplotlib==3.7.2
PyYAML==6.0.1
einops==0.6.1
compressai==1.2.4